+++ app/src/main/java/nl/hogeschoolrotterdam/projectb/MemoryDetailActivity.java
@@ -15,7 +15,7 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.core.content.ContextCompat;
 import androidx.viewpager2.widget.ViewPager2;
 import nl.hogeschoolrotterdam.projectb.data.Database;
-import nl.hogeschoolrotterdam.projectb.data.Memory;
+import nl.hogeschoolrotterdam.projectb.data.room.entities.Memory;
 
 public class MemoryDetailActivity extends AppCompatActivity {
     Intent shareIntent;
