+++ app/src/main/java/nl/hogeschoolrotterdam/projectb/MemoryEditActivity.java
@@ -17,8 +17,8 @@ import androidx.core.content.ContextCompat;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.material.textfield.TextInputLayout;
 import nl.hogeschoolrotterdam.projectb.data.Database;
-import nl.hogeschoolrotterdam.projectb.data.Memory;
-import nl.hogeschoolrotterdam.projectb.data.media.Media;
+import nl.hogeschoolrotterdam.projectb.data.room.entities.Media;
+import nl.hogeschoolrotterdam.projectb.data.room.entities.Memory;
 import nl.hogeschoolrotterdam.projectb.util.LocationManager;
 import nl.hogeschoolrotterdam.projectb.util.SimpleTextWatcher;
