--- app/src/main/java/nl/hogeschoolrotterdam/projectb/data/media/Media.java
@@ -1,4 +1,0 @@
-package nl.hogeschoolrotterdam.projectb.data.media;
-
-public abstract class Media {
-}
