+++ app/src/main/java/nl/hogeschoolrotterdam/projectb/data/room/entities/Media.java
@@ -1,0 +1,4 @@
+package nl.hogeschoolrotterdam.projectb.data.room.entities;
+
+public abstract class Media {
+}
